Move Program.cs into Example
